requirements.txt
ultralytics
opencv-python
torch
torchvision
numpy
pyyaml
tqdm
matplotlib
pillow
scipy